feat(map): add an optional first location to map
It is now possible to pass a first location to the map screen, that will
be used in priority. This will be useful for location buttons linked to
challenges.
If no first location is defined, then the map screen will try to center
on the user's location, otherwise it will default to Nice, France.