fix: replace Timestamp by Date throughout the project
Removed Timestamp imports, usage and type usages through the whole project and replaced it by Date which is a more simpler way of dealing with dates. It was causing issues because of how we managed it and I realised we didn't need it since we could change directly the type when we fetch from firebase.