fix: remove faulty dup check
One method used to create a group checked some duplication that
used the GID of the group before it was sent to Firebase, where
the GID is set. This made it impossible to create a group. It was fixed
by removing the duplicate check, as 2 groups cannot have the same GID
anyway.